Fixing up some small code mishaps

+++ A1/baseline.c
@@ -11,7 +11,7 @@ int main(){
     clock_t cpu_time_used;
     start = clock();
     long long sum = 0;
-    for(int i = 0; i < N; i++){
+    for(long long i = 0; i < N; i++){
         sum += i;
     }
 
@@ -19,7 +19,7 @@ int main(){
     cpu_time_used = end - start;
     double cpu_seconds = (double)cpu_time_used/CLOCKS_PER_SEC;
 
-    printf("Total sum from 0 to %d is: %lld\n",N,sum);
+    printf("Total sum from 0 to %lld is: %lld\n",N,sum);
     printf("Total CPU time taken in seconds: %.15f\n",cpu_seconds);
 
     return 0;
